[#942] Cleanup of _class2_compute.py into _class02_sample.py and _class02_tangency_radius.py
