+++ tofu/data/_class02_Rays.py
@@ -5,15 +5,11 @@
 # import copy
 
 
-# Common
-import numpy as np
-import datastock as ds
-
-
 # tofu
 from ._class01_Plasma2D import Plasma2D as Previous
 from . import _class2_check as _check
-from . import _class2_compute as _compute
+from . import _class02_sample as _sample
+from . import _class02_tangency_radius as _tangency_radius
 from . import _class2_plot as _plot
 from . import _class2_sinogram as _sinogram
 from . import _class02_save2stp as _save2stp
@@ -192,7 +188,7 @@ class Rays(Previous):
             flag indicating whether to concatenate the sampled points per ray
         """
 
-        return _compute._sample(
+        return _sample.main(
             coll=self,
             key=key,
             key_cam=key_cam,
@@ -245,7 +241,7 @@ class Rays(Previous):
             The ref tuple on which the data depends
         """
 
-        return _compute._tangency_radius(
+        return _tangency_radius._tangency_radius(
             coll=self,
             key=key,
             key_cam=key_cam,
@@ -302,7 +298,7 @@ class Rays(Previous):
 
         """
 
-        return _compute.intersect_radius(
+        return _tangency_radius.intersect_radius(
             coll=self,
             key=key,
             key_cam=key_cam,
